pyproject.toml
[tool.poetry]
name = "qref"
version = "0.4.0"
description = "Quantum Resource Estimation Format"
authors = ["PsiQuantum, Corp."]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.0"
graphviz = "^0.20.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
mypy = "^1.9.0"
flake8 = "^7.0.0"
black = "^24.2.0"
pyyaml = "^6"
jsonschema = "^4"



[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-link-marker = "^0.1.3"
mkdocs-material = "^9.5.15"
mkdocs-open-in-new-tab = "^1.0.3"
mkdocstrings = {version = "^0.24.2", extras = ["python"]}
mike = "^2.0.0"

[tool.poetry.scripts]
qref-render = "qref.experimental.rendering:render_entry_point"

[tool.poetry-dynamic-versioning]
enable = true
pattern = "default-unprefixed"


[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"


[tool.black]
line-length = 120
target-version = ['py39']


[[tool.mypy.overrides]]
module = "graphviz.*"
ignore_missing_imports = true
